Disallowing the use of API Model.create and Model.remove (#12)
These two methods can't be protected by this plugin, so it's best to disallow them entirely. By doing so, it makes it safe to wrap your mongoose models with an Restify wrapper.

Also, this diff introduces Ava and the new testing engine. Very similar syntax as before, just a newer, modern, maintained library.