Updates to post save permissions updates
Previously, if you did a `find()` with permissions, then did a save on that doc with `{ authLevel: false }`, the resulting document would still have the original permissions and it would not reflect the latest data in the document. This change stores the query options used for the find query in the document at find time. Then, if you take an action that skips authz, the permissions in the doc will be recalculated using those original options.